+++ investops/__init__.py
@@ -1,2 +1,2 @@
 # The version is also defined in setup.py and must be updated in both places.
-__version__ = "0.1.0"
+__version__ = "0.2.0"